First attempt at espresso
