+++ .travis.yml
@@ -12,4 +12,4 @@ matrix:
   fast_finish: true
 script:
   - npx truffle compile
-  - node mocha.js
+  - npm test
