+++ test/bump.test.js
@@ -1,13 +1,13 @@
 'use strict'
-require('../setup')
+require('./setup')
 
-import init from '../../src/scripts/init.js';
-import initLib from '../../src/scripts/init-lib.js';
-import bumpVersion from '../../src/scripts/bump.js';
-import linkStdlib from '../../src/scripts/link.js';
+import init from '../src/scripts/init.js';
+import initLib from '../src/scripts/init-lib.js';
+import bumpVersion from '../src/scripts/bump.js';
+import linkStdlib from '../src/scripts/link.js';
 import { FileSystem as fs } from 'zos-lib';
-import { cleanup, cleanupfn } from '../helpers/cleanup.js';
-import add from '../../src/scripts/add.js';
+import { cleanup, cleanupfn } from './helpers/cleanup.js';
+import add from '../src/scripts/add.js';
 
 contract('new-version script', function() {
   const appName = 'MyApp';
