+++ test/mocks/packages/package-with-invalid-stdlib.zos.test.json
@@ -1,0 +1,4 @@
+{
+  "contracts": {},
+  "proxies": {}
+}
